incident_kg_project/requirements.txt
langchain
langchain-community
oci
neo4j
faker
python-dotenv
faiss-cpu
